refactor test suite
