+++ test/reconciler/mock/asgclient.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package test
+package mock
 
 import (
 	"github.com/aws/aws-sdk-go/aws"
