+++ test/reconciler/mock/ec2client.go
@@ -14,16 +14,20 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package test
+package mock
 
 import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/aws/request"
+	"github.com/aws/aws-sdk-go/service/ec2"
 )
 
-func TestAppIntegration(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "App Integration Suite")
+type (
+	DescribeEC2InstancesFunc = func(aws.Context, *ec2.DescribeInstancesInput, ...request.Option) (*ec2.DescribeInstancesOutput, error)
+
+	EC2Client DescribeEC2InstancesFunc
+)
+
+func (e EC2Client) DescribeInstancesWithContext(ctx aws.Context, input *ec2.DescribeInstancesInput, options ...request.Option) (*ec2.DescribeInstancesOutput, error) {
+	return e(ctx, input, options...)
 }
